condense train_utils.py into utils.py
